Create markers (door, story, room); triangulate
Add story-marker to make sense of level structure.
Can be a barrier (with linked collider to partition the room), a key or
a regular marker (just with index).
Add door-marker to be able to place hallway next to door; draw direction
handle to indicate the direction of the marker (relevant for placement
on corners).
Adding two new marker-types (with potentially linked
meshfilter-components) confuses the reading in of rooms, add a room
marker to clearly identify gameobjects, which are to be treated as
rooms.
Use Delaunay Triangulation to get a sense, which rooms are near each
other.